PathfindingAlgorithm added to agent

+++ src/Connection.cpp
@@ -5,7 +5,7 @@ Connection::Connection() {
 
 }
 
-Connection::Connection(float _cost, float _nodeFrom, float _nodeTo) {
+Connection::Connection(float _cost, int _nodeFrom, int _nodeTo) {
 	cost = _cost;
 	nodeFromID = _nodeFrom;
 	nodeToID = _nodeTo;
